# docs/swagger.yaml
basePath: /api
...
host: localhost:5000
info:
  contact: {}
  description: 兽剧内容聚合平台后端服务 API 文档（neo-server，行为对齐 Express 版）
  title: 兽剧聚合平台 API
  version: 1.0.0
paths:
  /2fa/disable:
    post:
      consumes:
      - application/json
      parameters:
      - description: TOTP 或备份码
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 关闭 2FA
      tags:
      - 2FA
  /2fa/enable:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: secret/backupCodes/otpauthUrl
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 开启 2FA（生成密钥）
      tags:
      - 2FA
  /2fa/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: TOTP 或备份码
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: verified
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - bearerAuth: []
      summary: 校验 2FA 码
      tags:
      - 2FA
  /2fa/verify-enable:
    post:
      consumes:
      - application/json
      parameters:
      - description: TOTP 码
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 确认开启 2FA
      tags:
      - 2FA
...
  /users/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片（≤2MB）
        in: formData
        name: avatar
        required: true
        type: file
      responses:
        "200":
          description: url
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 上传头像
      tags:
      - 用户
  /users/background-prefs:
    put:
      consumes:
      - application/json
      parameters:
      - description: enabled/opacity/blur/image
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: backgroundPrefs
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 更新背景偏好
      tags:
      - 用户
  /users/background-upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片（≤5MB）
        in: formData
        name: image
        required: true
        type: file
      responses:
        "200":
          description: url
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 上传背景图
      tags:
      - 用户
  /users/export-my-data:
    get:
      parameters:
      - description: json|csv
        in: query
        name: format
        type: string
      responses:
        "200":
          description: 文件下载
          schema:
            type: string
      security:
      - bearerAuth: []
      summary: 导出个人数据
      tags:
      - 用户
  /users/profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: username
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 更新昵称
      tags:
      - 用户
securityDefinitions:
  bearerAuth:
    description: 输入 JWT access token（不含 Bearer 前缀；亦可通过 accessToken cookie）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"